requirements.txt
Flask
gpiozero
RPi.GPIO
basic-pitch
mido
